add: UserClient interface and impl
