+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/service/SpendClient.java
@@ -1,0 +1,16 @@
+package guru.qa.niffler.service;
+
+import guru.qa.niffler.model.CategoryJson;
+import guru.qa.niffler.model.SpendJson;
+
+public interface SpendClient {
+
+    SpendJson createSpend(SpendJson spend);
+
+    CategoryJson createCategory(CategoryJson category);
+
+    CategoryJson updateCategory(CategoryJson category);
+
+    void removeCategory(CategoryJson category);
+
+}
