--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/service/SpendDbClient.java
@@ -1,38 +1,0 @@
-package guru.qa.niffler.service;
-
-import guru.qa.niffler.config.Config;
-import guru.qa.niffler.data.dao.CategoryDao;
-import guru.qa.niffler.data.dao.SpendDao;
-import guru.qa.niffler.data.dao.impl.CategoryDaoJdbc;
-import guru.qa.niffler.data.dao.impl.SpendDaoJdbc;
-import guru.qa.niffler.data.entity.spend.CategoryEntity;
-import guru.qa.niffler.data.entity.spend.SpendEntity;
-import guru.qa.niffler.data.tpl.JdbcTransactionTemplate;
-import guru.qa.niffler.model.SpendJson;
-
-
-public class SpendDbClient {
-
-  private static final Config CFG = Config.getInstance();
-
-  private final CategoryDao categoryDao = new CategoryDaoJdbc();
-  private final SpendDao spendDao = new SpendDaoJdbc();
-
-  private final JdbcTransactionTemplate jdbcTxTemplate = new JdbcTransactionTemplate(
-      CFG.spendJdbcUrl()
-  );
-
-  public SpendJson createSpend(SpendJson spend) {
-    return jdbcTxTemplate.execute(() -> {
-          SpendEntity spendEntity = SpendEntity.fromJson(spend);
-          if (spendEntity.getCategory().getId() == null) {
-            CategoryEntity categoryEntity = categoryDao.create(spendEntity.getCategory());
-            spendEntity.setCategory(categoryEntity);
-          }
-          return SpendJson.fromEntity(
-              spendDao.create(spendEntity)
-          );
-        }
-    );
-  }
-}
