+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/service/UsersClient.java
@@ -1,0 +1,15 @@
+package guru.qa.niffler.service;
+
+import guru.qa.niffler.model.UserJson;
+
+public interface UsersClient {
+
+    UserJson createUser(String username, String password);
+
+    void createIncomeInvitations(UserJson targetUser, int count);
+
+    void createOutcomeInvitations(UserJson targetUser, int count);
+
+    void createFriends(UserJson targetUser, int count);
+
+}
