--- niffler-e-2-e-tests/src/test/java/guru/qa/niffler/service/UsersDbClient.java
@@ -1,148 +1,0 @@
-package guru.qa.niffler.service;
-
-import guru.qa.niffler.config.Config;
-import guru.qa.niffler.data.entity.auth.AuthUserEntity;
-import guru.qa.niffler.data.entity.auth.Authority;
-import guru.qa.niffler.data.entity.auth.AuthorityEntity;
-import guru.qa.niffler.data.entity.userdata.UserEntity;
-import guru.qa.niffler.data.repository.AuthUserRepository;
-import guru.qa.niffler.data.repository.UserdataUserRepository;
-import guru.qa.niffler.data.repository.impl.AuthUserRepositoryHibernate;
-import guru.qa.niffler.data.repository.impl.UserdataUserRepositoryHibernate;
-import guru.qa.niffler.data.tpl.DataSources;
-import guru.qa.niffler.data.tpl.XaTransactionTemplate;
-import guru.qa.niffler.model.CurrencyValues;
-import guru.qa.niffler.model.UserJson;
-import org.springframework.jdbc.support.JdbcTransactionManager;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.transaction.support.TransactionTemplate;
-
-import java.util.Arrays;
-
-import static guru.qa.niffler.utils.RandomDataUtils.randomUsername;
-
-
-public class UsersDbClient {
-
-  private static final Config CFG = Config.getInstance();
-  private static final PasswordEncoder pe = PasswordEncoderFactories.createDelegatingPasswordEncoder();
-
-  private final AuthUserRepository authUserRepository = new AuthUserRepositoryHibernate();
-  private final UserdataUserRepository userdataUserRepository = new UserdataUserRepositoryHibernate();
-
-  private final TransactionTemplate txTemplate = new TransactionTemplate(
-      new JdbcTransactionManager(
-          DataSources.dataSource(CFG.authJdbcUrl())
-      )
-  );
-
-  private final XaTransactionTemplate xaTransactionTemplate = new XaTransactionTemplate(
-      CFG.authJdbcUrl(),
-      CFG.userdataJdbcUrl()
-  );
-
-  @SuppressWarnings("unchecked")
-  public UserJson createUser(String username, String password) {
-    return xaTransactionTemplate.execute(() -> {
-          AuthUserEntity authUser = getAuthUserEntity(username, password);
-          authUserRepository.create(authUser);
-          return UserJson.fromEntity(
-              userdataUserRepository.create(userEntity(username)),
-              null
-          );
-        }
-    );
-  }
-
-  @SuppressWarnings("unchecked")
-  public void addIncomeInvitation(UserJson targetUser, int count) {
-    if (count > 0) {
-      UserEntity targetEntity = userdataUserRepository.findById(
-          targetUser.id()
-      ).orElseThrow();
-
-      for (int i = 0; i < count; i++) {
-        xaTransactionTemplate.execute(() -> {
-              String username = randomUsername();
-              AuthUserEntity authUser = getAuthUserEntity(username, "12345");
-              authUserRepository.create(authUser);
-              UserEntity addressee = userdataUserRepository.create(userEntity(username));
-              userdataUserRepository.sendInvitation(addressee, targetEntity);
-              return null;
-            }
-        );
-      }
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  public void addOutcomeInvitation(UserJson targetUser, int count) {
-    if (count > 0) {
-      UserEntity targetEntity = userdataUserRepository.findById(
-          targetUser.id()
-      ).orElseThrow();
-
-      for (int i = 0; i < count; i++) {
-        xaTransactionTemplate.execute(() -> {
-              String username = randomUsername();
-              AuthUserEntity authUser = getAuthUserEntity(username, "12345");
-              authUserRepository.create(authUser);
-              UserEntity addressee = userdataUserRepository.create(userEntity(username));
-              userdataUserRepository.sendInvitation(targetEntity, addressee);
-              return null;
-            }
-        );
-      }
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  void addFriend(UserJson targetUser, int count) {
-      if (count > 0) {
-          UserEntity targetEntity = userdataUserRepository.findById(
-                  targetUser.id()
-          ).orElseThrow();
-
-          for (int i = 0; i < count; i++) {
-              xaTransactionTemplate.execute(() -> {
-                          String username = randomUsername();
-                          AuthUserEntity authUser = getAuthUserEntity(username, "12345");
-                          authUserRepository.create(authUser);
-                          UserEntity addressee = userdataUserRepository.create(userEntity(username));
-                          userdataUserRepository.addFriend(targetEntity, addressee);
-                          return null;
-                      }
-              );
-          }
-      }
-  }
-
-  private UserEntity userEntity(String username) {
-    UserEntity ue = new UserEntity();
-    ue.setUsername(username);
-    ue.setCurrency(CurrencyValues.RUB);
-    return ue;
-  }
-
-  private AuthUserEntity getAuthUserEntity(String username, String password) {
-    AuthUserEntity authUser = new AuthUserEntity();
-    authUser.setUsername(username);
-    authUser.setPassword(pe.encode(password));
-    authUser.setEnabled(true);
-    authUser.setAccountNonExpired(true);
-    authUser.setAccountNonLocked(true);
-    authUser.setCredentialsNonExpired(true);
-    authUser.setAuthorities(
-        Arrays.stream(Authority.values()).map(
-            e -> {
-              AuthorityEntity ae = new AuthorityEntity();
-              ae.setUser(authUser);
-              ae.setAuthority(e);
-              return ae;
-            }
-        ).toList()
-    );
-    return authUser;
-  }
-}
