+++ niffler-e-2-e-tests/src/test/resources/META-INF/persistence.xml
@@ -26,4 +26,16 @@
         </properties>
     </persistence-unit>
 
+    <persistence-unit name="jdbc:postgresql://127.0.0.1:5432/niffler-spend" transaction-type="JTA">
+        <jta-data-source>java:comp/env/jdbc/niffler-spend</jta-data-source>
+        <class>guru.qa.niffler.data.entity.spend.CategoryEntity</class>
+        <class>guru.qa.niffler.data.entity.spend.SpendEntity</class>
+
+        <properties>
+            <property name="hibernate.transaction.jta.platform" value="com.atomikos.icatch.jta.hibernate4.AtomikosPlatform"/>
+            <property name="hibernate.connection.handling_mode" value="DELAYED_ACQUISITION_AND_RELEASE_AFTER_STATEMENT"/>
+            <property name="hibernate.dialect" value="org.hibernate.dialect.PostgreSQLDialect"/>
+        </properties>
+    </persistence-unit>
+
 </persistence>
